move CryptoParser
