+++ crypto-scraper/app/Console/Commands/Bitcoinabuse/Parse.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace App\Console\Commands\Bitcoinabuse;
 
@@ -7,7 +8,7 @@ use App\Console\ParserInterface;
 use App\Console\Utils;
 use App\Models\ParsedAddress;
 use Symfony\Component\DomCrawler\Crawler;
-use App\Console\Commands\CryptoParser;
+use App\Console\CryptoParser;
 
 class Parse extends CryptoParser implements ParserInterface
 {
